requirements.txt
django
django-allauth
django-crispy-forms
crispy-tailwind
django-role-permissions
django-admin-interface
django-tailwind[reload]